move types in a separate file
