+++ src/types.ts
@@ -1,0 +1,11 @@
+import { ReplaceOperation, AddOperation } from 'fast-json-patch';
+
+export interface DiffOutput {
+  action: string;
+  path: string;
+  index?: number;
+  before?: any;
+  after?: any;
+}
+
+export type ValueOperation = ReplaceOperation<any> | AddOperation<any>;
